Tests for GetCreditsWithInvoices

+++ CreditModule.DomainModels/Invoice.cs
@@ -4,6 +4,6 @@
     {
         public int InvoiceId { get; set; }
         public int InvoiceNumber { get; set; }
-        public int InvoiceTotal { get; set; }
+        public decimal InvoiceTotal { get; set; }
     }
 }
